+++ app/jobs/articles/bust_cache_job.rb
@@ -2,11 +2,10 @@ module Articles
   class BustCacheJob < ApplicationJob
     queue_as :articles_bust_cache
 
-    def perform(article_ids, cache_buster = CacheBuster.new)
-      Article.select(:id, :path).where(id: article_ids).find_each do |article|
-        cache_buster.bust(article.path)
-        cache_buster.bust(article.path + "?i=i")
-      end
+    def perform(article_id, cache_buster = CacheBuster.new)
+      article = Article.find_by(id: article_id)
+
+      cache_buster.bust_article(article) if article
     end
   end
 end
